@@ -1,18 +1,20 @@
 {
   "name": "terminal-kit",
-  "version": "0.24.30",
+  "version": "0.25.0",
   "description": "256 colors, keys and mouse, input field, screen buffer, interactive 'yes or no', and many more... Whether you just need colors and styles, build a simple interactive command line tool or a complexe terminal app: this is the absolute terminal lib for Node.js!",
   "main": "lib/termkit.js",
   "directories": {
     "test": "test"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=4.5.0"
   },
   "dependencies": {
-    "async-kit": "^2.1.6",
-    "nextgen-events": "^0.9.5",
-    "string-kit": "^0.5.10",
+    "async-kit": "^2.2.0",
+    "jshint": "^2.9.4",
+    "mocha": "^3.1.2",
+    "nextgen-events": "^0.9.8",
+    "string-kit": "^0.5.11",
     "tree-kit": "^0.5.26"
   },
   "devDependencies": {
@@ -21,7 +23,7 @@
     "mocha": "^3.0.2"
   },
   "scripts": {
-    "test": "mocha -R dot"
+    "test": "tea-time -R dot"
   },
   "repository": {
     "type": "git",
