@@ -87,7 +87,7 @@ ScreenBuffer.create = function create( options )
 	
 	Object.defineProperties( screenBuffer , {
 		buffer: { enumerable: true , configurable: true ,
-			value: new Buffer( screenBuffer.width * screenBuffer.height * ScreenBuffer.ITEM_SIZE ) 
+			value: Buffer.allocUnsafe( screenBuffer.width * screenBuffer.height * ScreenBuffer.ITEM_SIZE ) 
 		}
 	} ) ;
 	
@@ -180,7 +180,7 @@ ScreenBuffer.prototype.fill = function fill( options )
 	
 	if ( options && typeof options === 'object' ) 
 	{
-		clearBuffer = new Buffer( ScreenBuffer.ITEM_SIZE ) ;
+		clearBuffer = Buffer.allocUnsafe( ScreenBuffer.ITEM_SIZE ) ;
 		
 		// Write the attributes
 		attr = options.attr !== undefined ? options.attr : ScreenBuffer.DEFAULT_ATTR ;
@@ -379,7 +379,7 @@ ScreenBuffer.prototype.resize = function resize( fromRect )
 	if ( toRect.isNull ) { return false ; }
 	
 	// Generate a new buffer
-	var resizedBuffer = new Buffer( toRect.width * toRect.height * ScreenBuffer.ITEM_SIZE ) ;
+	var resizedBuffer = Buffer.allocUnsafe( toRect.width * toRect.height * ScreenBuffer.ITEM_SIZE ) ;
 	this.fill( { buffer: resizedBuffer } ) ;
 	
 	// We use the blit to reconstruct the buffer geometry
@@ -622,7 +622,7 @@ ScreenBuffer.prototype.terminalBlitter = function terminalBlitter( p )
 	{
 		if ( ! this.lastBuffer || this.lastBuffer.length !== this.buffer.length )
 		{
-			this.lastBuffer = new Buffer( this.buffer ) ;
+			this.lastBuffer = Buffer.from( this.buffer ) ;
 		}
 		else if ( this.lastBufferUpToDate )
 		{
@@ -792,7 +792,7 @@ ScreenBuffer.prototype.saveSyncV1 = function saveSync( filepath )
 {
 	var content ;
 	
-	content = new Buffer( ScreenBuffer.HEADER_SIZE + this.buffer.length ) ;
+	content = Buffer.allocUnsafe( ScreenBuffer.HEADER_SIZE + this.buffer.length ) ;
 	
 	// Clear the header area
 	content.fill( 0 , 0 , ScreenBuffer.HEADER_SIZE ) ;
@@ -886,7 +886,7 @@ ScreenBuffer.prototype.saveSyncV2 = function saveSync( filepath )
 	
 	header = 'SB\n' + JSON.stringify( header ) + '\n' ;
 	
-	content = new Buffer( header.length + this.buffer.length ) ;
+	content = Buffer.allocUnsafe( header.length + this.buffer.length ) ;
 	content.write( header ) ;
 	
 	this.buffer.copy( content , header.length ) ;
@@ -1164,7 +1164,7 @@ ScreenBuffer.ITEM_SIZE = ScreenBuffer.ATTR_SIZE + ScreenBuffer.CHAR_SIZE ;
 
 ScreenBuffer.DEFAULT_ATTR = ScreenBuffer.object2attr( { color: 'white' , bgColor: 'black' } ) ;
 ScreenBuffer.CLEAR_ATTR = ScreenBuffer.object2attr( { color: 'white' , bgColor: 'black' , transparency: true } ) ;
-ScreenBuffer.CLEAR_BUFFER = new Buffer( ScreenBuffer.ITEM_SIZE ) ;
+ScreenBuffer.CLEAR_BUFFER = Buffer.allocUnsafe( ScreenBuffer.ITEM_SIZE ) ;
 ScreenBuffer.CLEAR_BUFFER.writeInt32BE( ScreenBuffer.DEFAULT_ATTR , 0 ) ;
 ScreenBuffer.CLEAR_BUFFER.write( ' \x00\x00\x00' , ScreenBuffer.ATTR_SIZE ) ;	// space
 
