@@ -56,7 +56,7 @@ module.exports = gpm ;
 
 
 // Can't figure out the usage of the GPM_MAGIC constant ATM
-var gpmMagic = new Buffer( 4 ) ;
+var gpmMagic = Buffer.allocUnsafe( 4 ) ;
 gpmMagic.writeUInt32LE( 0x47706D4C , 0 ) ;
 
 
@@ -64,7 +64,7 @@ gpmMagic.writeUInt32LE( 0x47706D4C , 0 ) ;
 // Return a Buffer containing a Gpm_Connect structure, using a pid and a ttyIndex
 gpm.connectStructureBuffer = function connectStructureBuffer( gpmConnect )
 {
-	var buffer = new Buffer( 16 ) ;
+	var buffer = Buffer.allocUnsafe( 16 ) ;
 	
 	if ( gpmConnect.eventMask === undefined ) { gpmConnect.eventMask = 65535 ; }
 	if ( gpmConnect.defaultMask === undefined ) { gpmConnect.defaultMask = 0 ; }
